Homepage changes per team discussion
