+++ src/App.css
@@ -28,7 +28,7 @@ textarea {
 
 .main-search-bar {
   display: inline-block;
-  padding-top: 8px;
+  padding-top: 7px;
   width: 192px;
 }
